feat(Makefile): add test target to run Go tests with verbose output
The test target was added to the Makefile to enable running Go tests easily using the 'make test' command. The target uses the 'go test' command with the -v flag to provide verbose output, showing all the tests being run and their results. This addition improves the development workflow by making it convenient to run tests and get detailed feedback.